stockHandler y stockController
