+++ src/routes/dishesRouter.js
@@ -7,14 +7,15 @@ const {
   getDetailHandler,
   deleteDishesHandler,
   putDishesHandler,
-  getStockHandler
 } = require('../handlers/dishesHandler');
+const { getStockHandler, putStockHandler } = require('../handlers/stockHandler');
 const { validateDish } = require('../utils');
 const checkJwt = require('../utils/auth');
 
 const dishesRouter = Router();
 
 dishesRouter.get('/:id/stock', getStockHandler);
+dishesRouter.put('/:id/stock', putStockHandler);
 dishesRouter.get('/:id', getDetailHandler);
 dishesRouter.get('/', getDishesHandler);
 dishesRouter.post('/', upload.array('images', 10), validateDish, createDishesHandler);
